requirements.txt
pyarrow
apache-airflow-providers-amazon
